# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Ignite Backend (Prototype)
  version: 0.2.0
  description: >
    Authentication endpoints for registration with OTP, verification, resend,
    passwordless resume tokens, login, and token refresh.
servers:
  - url: http://localhost:5000/api
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /profile/me:
    get:
      tags: ["Profile"]
      summary: Get current user's profile core details
      description: |
        Returns the full profile record for the authenticated user (owner view; no visibility filtering).
        This allows users to view their own profile information.
        
        **Auth:** Bearer token required.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetProfileOK"
              examples:
                complete_profile:
                  value:
                    id: "123e4567-e89b-12d3-a456-426614174000"
                    fullName: "Jane Doe"
                    handle: "janedoe"
                    photoUrl: "https://example.com/profile.jpg"
                    bannerUrl: "https://example.com/banner.jpg"
                    isIndonesian: true
                    program: "BE"
                    major: "Computer Science"
                    level: "undergrad"
                    yearStart: 2020
                    yearGrad: 2024
                    zid: "z1234567"
                    headline: "Software Engineer"
                    domicileCity: "Sydney"
                    domicileCountry: "AU"
                    citizenshipStatus: "Permanent Resident"
                    bio: "Passionate about building innovative software solutions."
                    socialLinks:
                      linkedin: "https://www.linkedin.com/in/janedoe"
                      github: "https://github.com/janedoe"
                      website: "https://janedoe.dev"
                    createdAt: "2024-01-15T10:30:00Z"
                    updatedAt: "2024-03-20T14:45:00Z"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                not_authenticated:
                  value:
                    code: NOT_AUTHENTICATED
  /profile:
    patch:
      tags: ["Profile"]
      summary: Update user profile basic details
      description: |
        Updates core profile fields for the authenticated user. Only provided fields are updated (partial update).
        Excludes photoUrl, handle, and socialLinks which have their own dedicated endpoints.
        
        **Validation:**
        - `level`: Must be one of foundation, diploma, undergrad, postgrad, phd
        - `yearStart`, `yearGrad`: Must be between 2000-2100
        - `yearGrad`: Must be after `yearStart` if both provided
        - `domicileCountry`: Must be ISO country code (2 uppercase letters)
        - `citizenshipStatus`: Must be either Citizen or Permanent Resident
        
        **Auth:** Bearer token required.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
            examples:
              full_update:
                summary: Complete profile update
                value:
                  fullName: "Jane Doe"
                  headline: "Backend Engineer @ Acme | AI/ML"
                  isIndonesian: true
                  program: "BE"
                  major: "Software Engineering"
                  level: "undergrad"
                  yearStart: 2022
                  yearGrad: 2026
                  domicileCity: "Sydney"
                  domicileCountry: "AU"
                  citizenshipStatus: "Citizen"
                  bio: "I ship boring, reliable services."
              partial_update:
                summary: Partial update (only some fields)
                value:
                  fullName: "Updated Name"
                  bio: "Updated bio only"
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessOnly"
              examples:
                success:
                  value:
                    success: true
        "400":
          description: Validation error; Invalid country code, invalid academic level enum, invalid year range
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                validation_error:
                  value:
                    code: VALIDATION_ERROR
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                not_authenticated:
                  value:
                    code: NOT_AUTHENTICATED
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                internal_error:
                  value:
                    code: INTERNAL 
  /profile/skills:
    get:
      tags: ["Profile"]
      summary: List all skills for the authenticated user
      description: |
        Returns all skills linked to the authenticated user's profile via `profile_skills`.
        Names are deduped and sorted alphabetically.
        
        **Auth:** Bearer token required.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of skills for the user (may be empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
              examples:
                some_skills:
                  value:
                    - id: 12
                      name: Python
                    - id: 34
                      name: SQL
                no_skills:
                  value: []
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                not_authenticated:
                  value:
                    code: NOT_AUTHENTICATED
                    details: "Invalid or expired token"
    post:
      tags: ["Profile"]
      summary: Add a skill to the authenticated user's profile
      description: |
        Adds a skill to the user's profile. The skill is normalized and ensured to exist in the skills lookup table. Association is idempotent.
        
        **Auth:** Bearer token required.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skill]
              properties:
                skill:
                  type: string
                  example: Python
      responses:
        "201":
          description: Skill added to profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  id:
                    type: integer
                    example: 12
                  name:
                    type: string
                    example: Python
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
              examples:
                validation_error:
                  value:
                    code: VALIDATION_ERROR
                    details:
                      skill: Skill is required
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                not_authenticated:
                  value:
                    code: NOT_AUTHENTICATED
  /profile/skills/{id}:
    delete:
      tags: ["Profile"]
      summary: Remove a skill from the authenticated user's profile
      description: |
        Deletes the association between the authenticated user and the given skill by id.
        Idempotent: returns 200 even if the association does not exist.
        
        **Auth:** Bearer token required.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Skill ID to remove from the profile
          schema:
            type: integer
      responses:
        "200":
          description: Skill association removed (or already deleted)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
              examples:
                deleted:
                  value:
                    success: true
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                not_authenticated:
                  value:
                    code: NOT_AUTHENTICATED
        "404":
          description: Skill not found or not owned by user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                not_found:
                  value:
                    code: NOT_FOUND
...
  /profile/social-links:
    patch:
      tags: ["Profile"]
      summary: Update social links
      description: Replace entire `social_links` JSON with a validated object.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [socialLinks]
              properties:
                socialLinks:
                  type: object
                  additionalProperties: false
                  properties:
                    linkedin: { type: string, format: uri, example: "https://www.linkedin.com/in/janedoe" }
                    github:   { type: string, format: uri, example: "https://github.com/janedoe" }
                    x:        { type: string, format: uri, example: "https://x.com/janedoe" }
                    website:  { type: string, format: uri, example: "https://janedoe.dev" }
      responses:
        "200":
          description: Replaced successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }
        "401":
          description: Not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorSimple" }
        "500":
          $ref: "#/components/responses/InternalError"
...
  /profile/handle:
    patch:
      tags: ["Handles"]
      summary: Set or change profile handle (slug)
      description: >
        Claim or update a unique, public handle for the authenticated user.
        Must match `^[a-z0-9_.-]{3,30}$` and is unique case-insensitively.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandleRequest"
            examples:
              ok: { value: { handle: "janedoe" } }
      responses:
        "200":
          description: Handle set successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HandleOK"
              examples:
                ok: { value: { success: true, handle: "janedoe" } }
        "400":
          description: Invalid format or length
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorSimple" }
              examples:
                bad_format: { value: { code: VALIDATION_ERROR } }
        "401":
          description: Not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorSimple" }
              examples:
                no_auth: { value: { code: NOT_AUTHENTICATED } }
        "409":
          description: Handle already taken
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorSimple" }
              examples:
                taken: { value: { code: HANDLE_TAKEN } }
        "500":
          $ref: "#/components/responses/InternalError"
  /handles/check:
    get:
      tags: ["Handles"]
      summary: Check handle availability
      description: >
        Validates format and checks if the handle is available (case-insensitive).
      parameters:
        - name: handle
          in: query
          required: true
          schema:
            type: string
            pattern: "^[a-z0-9_.-]{3,30}$"
          example: janedoe
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HandleAvailability"
              examples:
                free:  { value: { available: true } }
                taken: { value: { available: false } }
        "400":
          description: Invalid format
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorSimple" }
              examples:
                bad_format: { value: { code: VALIDATION_ERROR } }
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    InternalError:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorSimple"
          examples:
            internal: { value: { code: INTERNAL } }
...
  schemas:
...
    SuccessOnly:
      type: object
      required: [success]
      properties:
        success: { type: boolean, example: true }
...
    ErrorSimple:
      type: object
      required: [code]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - EMAIL_EXISTS
            - ZID_EXISTS
            - ZID_MISMATCH
            - PENDING_VERIFICATION_EXISTS
            - PENDING_NOT_FOUND
            - ALREADY_VERIFIED
            - RESUME_TOKEN_INVALID
            - OTP_INVALID
            - OTP_EXPIRED
            - OTP_LOCKED
            - OTP_COOLDOWN
            - OTP_RESEND_LIMIT
            - INVALID_CREDENTIALS
            - ACCOUNT_NOT_VERIFIED
            - NO_REFRESH_TOKEN
            - INVALID_REFRESH_TOKEN
            - NOT_AUTHENTICATED
            - USER_NOT_FOUND
            - NO_PENDING_EMAIL_CHANGE
            - TOO_MANY_REQUESTS
            - INTERNAL
            - HANDLE_TAKEN
            - FILE_TOO_LARGE              
            - UNSUPPORTED_MEDIA_TYPE      
            - INTERNAL_ERROR
            - NOT_FOUND
            - INVALID_STATE

        details:
          type: string
          description: Optional additional details about the error
...
    Skill:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
...